requirements.txt
# SPDX-FileCopyrightText: 2021 Bluesy1 <68259537+Bluesy1@users.noreply.github.com>
# SPDX-License-Identifier: MIT
aiohttp[speedups]==3.8.3
asyncpg==0.26.0
discord.py[speed] @ git+https://github.com/Rapptz/discord.py@03d7a9a7191f470356df1b4b375e8f2d858a575c#egg=discord.py
disrank==0.0.2
fluent.runtime == 0.3.1
jishaku @ git+https://github.com/Gorialis/jishaku@a2a3752e4f540b10a96b5c285771b1534e3040fa#egg=jishaku
orjson==3.8.0
pandas==1.5.0
Pillow==9.2.0
pytesseract==0.3.10
sentry-sdk==1.9.10
tomli == 2.0.1; python_version < "3.11"
typing-extensions == 4.4.0
urlextract==1.6.0
uvloop == 0.17.0
validators==0.20.0
